Brand-cajita.js and Q-caja.js both have minor edits to be consistent with more recent caja.js versions.
caja.js has been changed to introduce privileged fault handlers for
failed property access. By using this ourselves in caja.js, we can now
avoid replacing any of the primordial properties and methods. Instead,
we don't allow the ones we would have replaced (or wrapped), and
instead install a fault-handler for the inaccessible property. The
result is a bit slower for Caja code, but allows untranslated
JavaScript to proceed at full speed.