Submitting change mikesamuel/parse-tree-rewrite
This fixes a few problems with ExpressionSanitizerBaja, and I think
will make it
much easier to write parse tree modification passes in the future.  It
touches a large number of files but most of the changes are formulaic.

I noted a number of unittests were missing from AllTests, and when I
added them
back, found ExpressionSanitizerBaja failing assertions.  It's quite
likely that
it worked mostly ok without assertions turned on, but there is the potential
that nodes' children lists could be confused causing undefined behavior.

I fixed it and a number of less serious bugs by reworking AbstractParseTreeNode
in ways that Ihab suggested during the initial review.

Now instead of a Visitor being called with a ParseTreeNode, and being
able to
get the parent node and siblings by calling get{Parent,NextSibling,PrevSibling}
the visitor gets an AncestorChain which allows access to the same information.
And since a node can be a child of multiple parents, (though not the
same parent
more than once), we no longer run into the problems that ExpressionSanitizerBaja
did where it was adding a child to one node without removing it from
its parent.

Changes include:
- Removed getParent getNextSibling, and getPrevSibling methods so that a node
  can be in multiple parents' children lists.  This fixes a number of bugs in
  ExpressionSanitizerBaja that were not exposed until I ran the tests with
  assertions on.  The bugs were the result of adding a child to a parent
  without removing it from its old parent.
- Removed parentification and manual calling of childrenChanged.
- Defined an AncestorChain class that encapsulates the node and its parent...
- Modified the Visitor interface and ParseTreeNode's pre-order and post-order
  traversals to take an AncestorChain
- Made AbstractParseTreeNode's children list private and modified all derived
  classes to initialize their children via createMutation.  Added appendChild
  convenience method.
- Deleted acceptBreadthFirst, which was not used, and rewrote acceptPreOrder and
  acceptPostOrder to not depend on children chains.  To do this efficiently, I
  had to allow for replacement of the child list on mutation.  Adding a child
  has no ill effects, so there's no changing of child lists there, but on
  insertion and deletion elsewhere, I copy on write.  This makes initialization
  efficient, and allows for efficient traversal that is robust over changes to
  the tree by the visitor.
- Fixed bug: post-order check that prevents a node from being visited if it
  was removed from the tree as a result of processing the child was always
  causing the root to be skipped.
- Rewrote AbstractParseTreeNode's mutators.  As a side-effect of the removal
  of the sibling list and parent chain, we no longer need to require that
  children be AbstractParseTreeNodes.
- Upaded all clients of ParseTreeNode and Visitor to use AncestorChain instead
  of getParent,getPrevSibling,getNextSibling, and to use children() instead of
  the now private children field.
- Fixed bug: HtmlCompiler using HTML escaping scheme instead of HTML attribute
  escaping.