Implements gadget rewriting as Ihab speced at http://google-caja.googlecode.com/svn/trunk/doc/html/cajaOpenSocialGadgetRewriting/index.html
  It'll probably be easiest to review these files first:
- javatests/.../opensocial/example.xml & example-rewritten.xml
- java/.../opensocial/DefaultGadgetRewriter.java
- java/.../plugin/HtmlPluginCompiler.java
- java/.../plugin/HtmlCompiler.java



  Changed GadgetRewriter and UriReference to take an
ExternalReference instead of a URI so that if fetching or
rewriting fails, we can tell where the problem occurred.

  Changed UriCallback::retrieve to return a Reader instead
of an InputStream since UriCallback might have access to
HTTP response headers which specify a charset and is just
as capable of guessing a charset  as any of its clients.

  Changed DefaultGadgetRewriter to keep track of file
positions.

  Moved html parsing out of HtmlPluginCompiler and into
clients.  Added some convenience methods to DomParser to
simplify this.  The clients should have control over
whether the input is HTML or XHTML, and passing strings
around loses FilePosition info.

  Removed parseSpec method from HtmlPluginCompiler, and
made the constructor side-effectless.  Having work done in
the constructor was causing error messages to be lost when
the default message queue would be overwritten by a client
calling setMessageQueue.

  Added code to HtmlPluginCompiler to handle <script src=>
and <link rel=stylesheet>.

  Rewrote HtmlPluginCompiler and HtmlCompiler to make
example.xml pass and added some attribute policy support to
HtmlWhitelist to filter link targets.

  Cleaned up success checking in HtmlPluginCompiler.  Now
progress proceeds as long as there are no fatal errors, but
run signals failure if there are either errors or fatal
errors.

  Modified JsWriter to allow us to emit html chunks in ways
that don't involve just pushing them onto an array.

  Modified HtmlLexer to treat style tags content as CDATA
when in HTML mode.

  Moved CSS and HTML/XML escaping into Escaping.java, and
fixed a few bugs with XML and CSS escaping in the process.
Code that accumulated CSS onto a buffer might not follow an
escaped character with a space when it needed to.

  Consolidated duplicated MessageType code and added an
OpenSocialMessageType class.

  Deleted HtmlDomParser and added a method to DomParser to
parse a fragment of HTML or XML.  HTMLPluginCompiler was
embedding content in <div>...</div> which was screwing up
FilePositions and introducing a tag not in the spec.

  Implemented rendering for DomTree nodes.

  Moved unique identifier generation into PluginMeta.

  Added javatests/.../opensocial/example.xml which demos
handling of external styles, rewriting of URLs, and html
emitting.

  Fleshed out DefaultGadgetRewriterTest.