Fix PluginCompilerMain to work with Caja.
I need to do this so that I have something that can cajole the Domita unittests written in caja.

PluginCompiler currently only works with AAJA and BAJA and this fixes it to work with CAJA.  This removes quite a bit of the code that distinguishes it from HtmlPluginCompilerMain, and fixes a few minor CAJA bugs.