*mikesamuel/calendar-demo-rrules-22-Jan-2008@533 | mikesamuel | 2008-02-06 12:50:01 -0800 (Wed, 06 Feb 2008)
Description:

The second part of the calendar demo app.

See package.html for a description of the overall purpose of the demo.

Most of this is a JS port of code.google.com/p/google-rfc-2445 which
is Apache licensed.

I made a few changes to RhinoTestBed to allow me to write my unittests
in an html file which can be executed in a browser, but also easily
executed in Rhino in the regular test engine.