Third part of calendar demo.
This defines a LayoutPolicy that maps date ranges to pairs of grid
coordinates, and a Layout class that breaks events into chips which
are positioned on the grid using a LayoutPolicy.

The overlap file defines function that take event chips (the graphical
representation of part of an event) and that minimizes overlap by
reducing chips along their minor axis.

The last part requires a bitset implementation which should provide
more optimization fodder.