*mikesamuel/domita-first-pass-3-Feb-2008@582 | mikesamuel | 2008-02-14 13:09:10 -0800 (Thu, 14 Feb 2008)
Description:

First pass at DOMita.

DOMita is a minimalist tamed DOM, and is meant as a stopgap measure
until Mark S's tamed BOM is ready.




I rewrote the html-sanitizer.  It is now built around a SAX like
parser that published events like (startTag, tagName, attributes),
(closeTag, tagName), (pcdata, text).

I then moved the HTML element and attribute whitelists into
html4-defs.js.

I then added a file domita.js which defines two functions:
    attachDocumentStub(idPrefix, uriCallback, outers)
    plugin_dispatchEvent___(node, event, pluginId, handler)
The first attaches a tame document object to outers, and the second
receives and dispatches events safely.


===
API
===
You can grep domita for allowMethod and exportFields to get a list of
available fields and methods.  As of right now, they are:

    ___.allowMethod, TameNode,
       ['getNodeType', 'getNodeValue', 'getNodeName',
        'appendChild', 'insertBefore', 'removeChild', 'replaceChild',
        'getFirstChild', 'getLastChild', 'getNextSibling', 'getPrevSibling',
        'getElementsByTagName']);

    exportFields(TameNode,
       ['nodeType', 'nodeValue', 'nodeName', 'firstChild',
        'lastChild', 'nextSibling', 'prevSibling']);

    ___.allowMethod, TameTextNode,
       ['setNodeValue', 'getData', 'setData']);
    exportFields(TameTextNode,
       ['nodeValue', 'data']);

    ___.allowMethod, TameElement,
       ['addEventListener', 'getAttribute', 'setAttribute',
        'getClassName', 'setClassName', 'getId', 'setId',
        'getInnerHTML', 'setInnerHTML', 'getTagName']);
    exportFields(TameElement,
       ['className', 'id', 'innerHTML', 'tagName']);

    ___.allowMethod, TameFormElement,
       ['getElements']);
    exportFields(TameFormElement,
       ['elements']);

    ___.allowMethod, TameInputElement,
       ['getValue', 'setValue', 'focus', 'getForm']);
    exportFields(TameInputElement,
       ['value', 'form']);

    ___.allowMethod, TameEvent,
       ['getType', 'getTarget', 'getPageX', 'getPageY',
        'stopPropagation', 'getAltKey', 'getCtrlKey',
        'getMetaKey', 'getShiftKey', 'getButton', 'getClientX',
        'getClientY', 'getKeyCode', 'getWhich']);
    exportFields(TameEvent,
       ['type', 'target', 'pageX', 'pageY', 'altKey',
        'ctrlKey', 'metaKey', 'shiftKey', 'button',
        'clientX', 'clientY', 'keyCode', 'which']);

    ___.allowMethod, TameDocument,
       ['createElement', 'createTextNode', 'getElementById']);


===================
Authorities Granted
===================
Granting access to a Node grants access to the entire subtree.
Assuming the node can be edited, the entire subtree (except the root)
or portions can be deleted or replaced.  Attributes can be modified
subject to certain policy constraints.  innerHTML can be fetched and
can be set, but is run through a sanitizer.

Granting access to a document allows lookup of any node with an id
prefixed with the document's idPrefix, though the prefix is not
visible to clients.

Granting access to a document or node does not grant authority to
navigate any frame.


==============
Event Handling
==============
The html sanitizer does recognize and rewrite some event handlers, and
TameElement includes both attachEvent and addEventListener methods.

If innerHTML contains an event handler, it is matched against
   [return ]<functionName>([this[, event]])
The event handler can be a function name that optionally takes two
arguments, the node, and the event, in that order.

The function name is resolved in the gadget's scope, so a gadget can do
  function foo(nodeClicked,.clickEvent) { ... }
  node.innerHTML = '<a onclick="return foo(this, event)">click</a>'
and have it behave as expected.

Once we have a client side Cajita compiler, we can expand this to allow
arbitrary cajita.



==========
Still TODO
==========
The most glaring hole in this DOM implementation is the style
attribute.  The style attribute cannot currently be set at all since I
have no CSS sanitizer.  I can work on that next, or I can try and
resurrect the CSS templates, or I can try and implement the new style
partitioning scheme.

I have run the attached testcases on Firefox, but have not done
testing on other browsers.  I have also not written a lot of negative
tests.  Hopefully Peter can help with that.

The tame node and document implementations all have an "editable" flag
so that a client can delegate a read only capability to a document or
DOM sub-tree.  This is not exposed anywhere in the API.

setTimeout & setInterval, and the clearing equivalents.