*mikesamuel/retire-aaja-y-baja-10-Feb-2008@586 | mikesamuel | 2008-02-14 17:22:12 -0800 (Thu, 14 Feb 2008)
Description:



Gets rid of all the obsolete AAJA and BAJA code, and migrates the
unittests for those to Caja.  This also resurrects the CSS templates
scheme which will probably be needed for domita soon, and which I want
to demo to the open social templates people.

As a result of the ?AJA -> CAJA transition, this removes most of the
fields in PluginMeta, but curiously leaves the idPrefix and URI base
path, which are also obsolete.  The idPrefix is obsoleted by Mark's
getElementById scheme, but is still needed for now.  The uri prefix
is obsoleted by the URI callback, but I will fix that in another CL.

- Modified DefaultCajaRewriter to pass ArrayConstructors.

- Removed namespaceName, namespacePrivateName, rootDivId from
  PluginMeta.  namespaceName has been replaced with
  ReservedNames.OUTERS since the caja rewriter does not pay attention
  to PluginMeta.

- Ripped out most of plugin-base.js -- the old version of caja.js based on
  wrap-capability.js.  I will retire both of those soon, but they're still
  used by the end-to-end plugin tests, so I want to keep them around a bit
  longer.

- Deleted ExpressionSanitizer and ExpressionSanitizerBaja.  There is only caja!

- Removed the root DIV from the DefaultGadgetRewriter.

- Removed the --scheme flag from GadgetRewriterMain

- Rewrote the GXP and CSS template compilers to be Caja friendly.

- Rewrote PluginCompiler to be pipelined.  In 1 more CL I will be able to get
  rid of one of (HtmlPluginCompiler, PluginCompiler), and maybe one of
  (GxpCompiler, HtmlCompiler).

- Rewrote ExpressionSanitizerTest to use ExpressionSanitizerCaja.
  This is the only test that fails, and I think there are a few
  genuine errors in there.