Rewrite parse tree rendering to allow rendering of original source code next to rewritten source code.
This is meant to make it easier to debug caja code.

When stepping through rewritten Cajoled output, the user will see
the original source code in a comment to the left:
  /* mySource(); */     ___.asSimpleFunc(___OUTERS___.mySource)();

See SideBySideRendererTest.java and the supporting files in
javatests/com/google/caja/render.

- Defined TokenConsumer and rewrote all ParseTreeNode.render to dump
  tokens to RenderContext.
- Implemented a TokenConsumer that pretty prints javascript.
- Added stochastic tests for tokenization which shook out several bugs
  in javascript lexing &| rendering:
  - Removed codepoint 0x85 from list of newline codepoints in JSLexer
  - Fixed bug: JSLexer expected an exponent after a hex token that ends
    with 'e'
  - Fixed bug: regexp normalization was not balancing brackets.  Slashes
    inside square brackets do not end regexps.

TODO: hook this into the main classes.