To address bug 67, we need to treat NaN as a regular identifier.
This changes NaN, and Infinity which had been lexed as floating point
tokens, and now treats them as non-reserved words.

Also updates the renderer to render numeric literals properly.  If we
fold constants and get a numeric value of NaN or Infinity, we will now
render that parse tree in a manner that is independent of environment.