From http://code.google.com/p/google-caja/issues/detail?id=68
Parser does not accept a property as the key variable in a foreach, despite
the fact that JavaScript (at least Firefox) permits it. An example and the
stacktrace are pasted below. ECMA-262 section 12.6.4 specifies the syntax as --

  for ( <LeftHandSideExpression> in <Expression> ) <Statement>

     *   *   *   *   *

Parsing: for (a.k in x) { k; }

com.google.caja.lexer.ParseException: source:1+14 - 15: Expected ; not )
	at com.google.caja.lexer.TokenQueue.expectToken(TokenQueue.java:251)
	at com.google.caja.lexer.JsTokenQueue.expectToken(JsTokenQueue.java:69)
	at com.google.caja.parser.js.Parser.parseLoopOrSwitch(Parser.java:376)
	at
com.google.caja.parser.js.Parser.parseStatementWithoutLabel(Parser.java:481)
	at com.google.caja.parser.js.Parser.parseStatement(Parser.java:310)
	at com.google.caja.parser.js.Parser.parseTerminatedStatement(Parser.java:1175)
	at com.google.caja.parser.js.Parser.parse(Parser.java:266)
        ...


- Fixed parser
- Modified ForEachLoop.java to allow arbitrary lvalues as the first child.