Fixes bug 114: Try to parse "function(){..}" and you get a NullPointerException.
Also fixes a missing dependency in the build file on parser/js.

The punctuation trie was being used incorrectly in InputElementJoiner,
and was joining ..} into one token.

This only occurred for .. since that is the only string that has a
valid punctuation string as a prefix, and is a prefix of a valid
punctuation string, but is not itself a valid punctuation string.