*mikesamuel/fix-bug-134-onload-handlers@879 | mikesamuel | 2008-03-27 13:01:39 -0800 (Thu, 27 Mar 2008)
Description:

Fix bug 134:
Extract <body onload> handler and run it after all other gadget script.

Some gadgets use the onload attribute of body to execute
initialization script.

We disallow <body> tags since HTML 5 specifies that extra bodies
attributes are folded into the global body element so have global
effect.

Right now the HtmlSanitizer ignores a body without attributes, so if
we remove the onload, then we can allow the gadget below to function
as the author intended:

  <body onload="init()">
    <script>function init() { console.log('hello world'); }</script>
  </body>



This changes fixes RewriteHtmlStage to create a fake script tag just
before the </body> containing the text of the onload handler.  It does
not attempt to create an onload event, or to run the script in the
context of the body element.