*mikesamuel/bug-98-gadget-file-positions@874 | mikesamuel | 2008-03-26 22:54:10 -0800 (Wed, 26 Mar 2008)
Description:

Fixed bug 98: Now Gadget Parser uses our XML parser instead of the DOM
one to parse gadgets so we have accurate file positions into the XML
file when we calculate error message snippets.

Previously we assumed the start of the content element was at
character 1 of line 1 which was never correct.