Fixes bug 148: Treat BOM as whitespace.
This does not attempt to add uniform BOM handling to CharProducer
since that would not handle BOMs except at the beginning of files, and
would break char-in-file position counting.

Reading BOMs is part of the heuristics used by many languages to
settle on an encoding given an octet stream ( see
http://www.w3.org/TR/2004/CR-CSS21-20040225/syndata.html#q23 ) and so
BOM handling is done non-destructively on the byte stream and then the
the interpretation of the BOM character is allowed for in the lexical
grammer.

- Add codepoint \uFEFF to JsLexer.isSpaceChar.
- Added tests to JsLexerTest
- Did not change CssLexer or HtmlLexer.