Fix bug 149: fix CSS background property signature
background:#F7F7F7 url(/images/foo.gif) no-repeat scroll left top;
results in an error message:
   css property background has bad value
but it looks fine.




The rule was fine, as was all the backtracking code, except for some
special case handling.

A [ a || b ] has different semantics inside a repeating section than
normally since a [ a || b ]* allows each of a and b to appear at most
once.

The special handling for this was failing when the repeating section
was processed last as part of a larger [ a || b ] section.



- Fixed handling of ExclusiveSet signatures inside a Repeatition signature
- Added tests for background and background-position properties to
  CssValidatorTest.