A cut at {delete,canDelete}{Pub,Prop} in caja.js.
I don't know whether this is the right thing to do, because I don't know how deletion works with whitelisting, or whether a field that was settable and has been deleted should still be settable.

(1) Does not allow deletion on frozen objects, or deletion of foo__ fields.
(2) Allows deletion of o.foo and o.foo_ on JSON objects
(3) Allows deletion of o.foo but not o.foo_ on non-JSON objects
    unless ___.allowDelete(o, 'foo')
(4) Allows deletion of this.foo and this.foo_ from methods.