Modified cajoling of new, instanceof, and delete operators.
new did not work with first class constructors.
Specifically new foo.bar() failed.

delete was returning the wrong value -- it should return a boolean.
(delete foo) was failing to cajole.
(delete 4) was failing with a ClassCastException.
(delete this.foo) was failing when this refers to ___OUTERS___.

instanceof did not work with first class constructors.
(new foo.bar() instanceof foo.bar) failed to cajole.

- modified newCalllessCtor to cajole (new Date) by first recursively
  cajoling (new Date())
- modified newCtor to allow @ctor to be an atribtrary expression.
- removed obsolete rules newFunc and newBadCtor
- added return statements to deleteProp and deletePub in caja.js
- modified deleteProp to cajole the @this so that it is turned into
  t___ or ___OUTERS___ as appropriate.
- added rule deleteGlobal to handle (delete @v)
- added rule deleteNonLValue to handle (delete 4).
- modified rule otherInstanceof to allow arbitrary rhs.
- removed rule otherBadInstanceof