*mikesamuel/consolidate-parser-utility-fns@1112 | mikesamuel | 2008-04-15 13:49:06 -0800 (Tue, 15 Apr 2008)
Description:

Move all the duplicate parsing code from our junit tests into a base
class CajaTestCase which provides utility parsing and rendering
functions.

The testcase maintains a MessageQueue, MessageContext, and InputSource
for teach test run.  The InputSource, which is used when content is
parsed from a string literal in a unittest includes
TestCase.getName(), so STDERR/STDOUT logs are annotated with the test
that was run.