Fixes parser break on unexpected absence of an expected token:
Eg:  do while(true);

The while() gets interpreted as the body of the do loop
(ie. while(true){}) and the terminating while() of the outer loop is
missing.  Instead of reporting this, the parser throws a null pointer
exception.

Fixed by reporting the expected token and position the token was
expected.