This is part of a larger set of changes to make it easy to use Eclipse to develop and debug gadgets.
This CL adds the lolcat example as a sample within the caja plugin. An eclipse user can import lolcat as a project from a dropdown box and start developing using it as a base without having to muck with working out what files go where.  It also allows the user to preview the cajoled sample app.