A bug was causing us to incorrectly reject CSS <lengths> without a sign.
"left: 10px" was being rejected, whereas "left: +10px" was fine.

I added unittests, but a Rhino bug caused them to fail.  I filed that as
    https://bugzilla.mozilla.org/show_bug.cgi?id=448443
but the workaround was to remove the extra parentheses.

I also update rhino/js.jar to version 1.7 to see if that bug was fixed
in a new version, but it wasn't.