Simplify how synthetics are used in preparation for implementing Ihab's custom node idea for pluggable code generators.
Goal
====
Ihab pointed out that the current synthetic mechanism combined with
pipelining makes it impossible to make the TCB smaller than all
possible code that a pipeline author could write.

There are a couple existing pipeline stages that generate synthetic
content -- GXP templates for HTML, CSS templates, the HtmlCompiler
code that extracts HTML event handlers.

The goal of this CL is to lay the groundwork for a narrow definition
of synthetic, so that we can have both an extensible pipeline and a
small TCB.


Design
======
Create a few custom nodes with rigorous consistency checks that are
passed through by the DefaultCajaRewriter.

The custom nodes will be part of the TCB, and any user code they
contain must be exposed as children.  At render time, the custom nodes
construct a parse tree of regular node, and delegate rendering to
that.


Material Changes in this CL
===========================
Previously, any node could be marked synthetic.  Now only
FunctionConstructors and Identifiers can.  Since SyntheticNodes.java
now knows about JS node types, I moved it into ...parser.js from
...parser.

Previously there was one rewriter rule that did an expandAll on any
synthetic node.  Now there is one per specific construct that we want
to pass through.

Changed QuasiBuilder so that when it parses a quasiliteral,
identifiers that end in __ are marked synthetic since they are clearly
intentional rules violations.


Other Changes in this CL
========================
Got rid of all the semicolon insertion warnings from
DefaultCajaRewriterTest's logs.

Fixed bug:  parser was expecting a semicolon after with-statements.


Open Questions
==============
Why do we need to ignore synthetic functions in determining in which
scope tempraries are declared?


Still TODO
==========
Define custom nodes for html event handlers and other code generators,
and get rid of the rules for Identifier and FunctionConstructor.