Fix bug 391: Malformed HTML attributes handled incorrectly
http://code.google.com/p/google-caja/issues/detail?id=391

Invalid HTML causes Caja lexer to generate incorrect token stream for subsequent tags

What steps will reproduce the problem?
1. Feeding 
<div id='search' style=width: 565px; position: relative'><div id="x">
into the Caja lexer will cause it to start generating incorrect tokens. The style attribute without 
correct quoting causes all subsequent tokens to be incorrectly recognized. This issue was found 
during development of the HTML rewriter for Shindig. 

What is the expected output? What do you see instead?
The lexer should terminate the attribute when it sees the > which seems to be consistent with 
browser behavior




Changed the lexer in HTML mode, to treat subsequent tokens inside a
tag body as part of the preceding attribute value unless they are the
start of a tag name.

Changed DomNode.Value to work properly on unbalanced quotes in a value.

This means that far fewer bare words in tag bodies will be treated as
valueless attributes.