Bug 560: tameNodeList gives access to a bare function Bug 568: Domita: TameNodes need to consistently check editable
Made node list and item method frozen, changed item method to use
nodes instead of this.

Added editable___ checks to removeChild, appendChild, insertBefore.

Modified domita_test.js to test for failures in a way that doesn't
lead to spurious success and modified the styles so that failed tests
are visually obvious in the HTML page.