*mikesamuel/unrepresentable-int-fix@2145 | mikesamuel | 2008-08-09 18:05:59 -0800 (Sat, 09 Aug 2008)
Description:

Fixes a bug with unrepresentable integers.

Our parser warns when a hex literal is too large to be represented
in javascript, but fails to construct an appropriate node.

Previously, it just passed the unrepresentable number, as a long, to
IntegerLiteral, which has its own consistency checks, and so threw an
IllegalArgumentException.

This changes the Parser to construct a RealLiteral from the long
downcast to a double.  The strictfp keyword makes sure that the JVM
will not use an 80b FP register on x86, and so the RealLiteral will be
the same that would be produced by SpiderMonkey.