*mikesamuel/regexp-div-ambiguity-fixes@2190 | mikesamuel | 2008-08-14 21:12:22 -0800 (Thu, 14 Aug 2008)
Description:

Implemented David-Sarah's suggested changes to lexing around '/'.

One of the tests in JsLexerTest is disabled by an if (false).

This is a known error that does not introduce a security vulnerability
since we do not require that our lexing is perfect -- we require that
our rendered output will be interpreted correctly.