*mikesamuel/bug-730@2436 | mikesamuel | 2008-09-29 11:22:25 -0800 (Mon, 29 Sep 2008)
Description:

Issue 730:	 IE7 is cranky about <option> in .innerHTML

Modified DomAttributeEvents to check the depth of an optimization.





From the bug:

this fails to show any options when cajoled:

  <form><select id='x'><option>abc</option></select></form>

in contrast, this does work:

  <form><select><option>abc</option></select></form>

the problem is, version 1 cajoles to

   .ih('<option>abc</option>')

and ie7 strips out the leading option tag there,
maybe because it thinks that option is invalid in that context.
version 2 cajoles to

   .ih('<select><option>abc</option></select>')

which works fine.

example (run this without cajoling):

<html>
<body>
<form id="f1"></form>
<form id="f2"><select id="s2"></select></form>
<script>
  var f1 = document.getElementById('f1');
  f1.innerHTML = '<select><option>abc</option></select>';
  alert('f1 is ' + f1.innerHTML);

  var f2 = document.getElementById('f2');
  f2.s2.innerHTML = '<option>abc</option>';
  alert('f2 is ' + f2.innerHTML);
</script>
</body>
</html>