*mikesamuel/bug-589@2455 | mikesamuel | 2008-10-01 08:58:11 -0800 (Wed, 01 Oct 2008)
Description:

Issue 589: tameNode should not return a functional node for an unrecognized or unsafe element

tameNode will return a TamedElement for any element regardless of whether
the element name is recognized and safe.

If it isn't, we still need to return a node so that tree traversal works,
but should not return an editable node, or one that exposes details of the
node.