*mikesamuel/fix-valija-typeof@2442 | mikesamuel | 2008-10-01 00:21:05 -0800 (Wed, 01 Oct 2008)
Description:

Fix typeof under valija.

Previously
  typeof noSuchVarDeclared
threw an Error since readOuter from valija-cajita throws an Error when
there is no such variable in outers.

I made a copy of readOuterSilent that returns undefined instead of
throwing an Error, added an abbreviation '$v.ros' to the set of
symbols exported from valija-cajita.js and to the PermitTemplate, and
added a rule to DefaultValijaRewriter that matches typeof operators
where the operand is a Reference to an outer.

To test, I copied the typeof assertConsistent tests from
CajitaRewriterTest into a test in CommonJsRewriterTestCase that will
be run both for Cajita and Valija tests.