*mikesamuel/bug-731@2521 | mikesamuel | 2008-10-08 19:58:33 -0800 (Wed, 08 Oct 2008)
Description:

Issue 731:  IE7 doesn't let you change <input> type after the UI is rendered

this fails when cajoled in IE7:

  <form><input id='x' type='radio'></form>

the problem is, this sequence fails (uncajoled):

  <form id='a'></form>
  <script>
    var a = document.getElementById('a');
    var i = document.createElement('input');
    a.appendChild(i);                   // line A
    i.setAttribute('type', 'radio');    // line B
  </script>

but it works if the order of line A and line B is reversed.

so I think a simple fix would be to have html-emitter defer attaching the
dom nodes to the real document, perhaps in f(), or perhaps later.