Partially addresses bug-614.
I thought I'd write down some of the ideas we had for trying to confuse the
runtime HTML sanitizer in Domita into allowing us to put in blacklisted
content (e.g., scripts).  These are ideas we brainstormed but ran out of
time to try/test.

1. SGML SHORTTAGs are fun.  You might not expect it, but the following
could possibly be interpreted as HTML markup (they might all mean the same
thing):

<FOO><BAR>
<FOO/BAR/
<FOO<BAR>

...

7. Try putting UTF-7 in the attribute.  Can you introduce the UTF-7
encoding of a double-quote?  What can you do with that?

8. Try slashes (see SGML shorttags).