*mikesamuel/lint-character-escapes@2583 | mikesamuel | 2008-10-14 17:00:53 -0800 (Tue, 14 Oct 2008)
Description:

Add a lint check to warn when a javascript string literal contains an
escaped character that has no special meaning in a string, but does in
a regular expression.

This should help catch bugs such as the one in html-sanitizer where a
RegExp read new RegExp('...\s*...') instead of new
RegExp('...\\s*...').