Fixes bugs 641, 422, 802, 803, 418, 806, and 799.
       641: typeof vs regexp objects

A reliable typeof for use by Cajita code, and by uncajoled code (like
parts of cajita.js) that require a reliable typeof.

ES3 and forward specify that <tt>typeof new RegExp("x")</tt> evaluate
to <tt>'object'</tt>. Unfortunately, on some of Cajita's current
target platforms (including at least Safari 3 and Rhino), it returns
<tt>'function'</tt> instead. Since the distinction between functions
and non-functions is crucial to Cajita, we translate the Cajita
<tt>typeof</tt> operator into calls to this <tt>typeOf</tt> function.

    422: makeSealerUnsealerPair() forgot to freeze its seal and unseal
    functions.

These are now marked as simple frozen functions.


    802: Fix cajita.getOwnPropertyNames(obj)

The old one had two problems: 1) it was now filtering out inherited
properties, and 2) it was including only properties explicitly
whilelisted, ignoring own properties on JSON containers that are
implicitly accessible because the underlying object is a JSON
container. Both these problems are now fixed.


    803, 418, 806: What is '[].push' in Cajita?

Since the tamed push must still enable '[].push(3)' to operate on its
receiver (the array on the left that becomes bound to push's
'this'-value, it is unclear what '[].push' should be. The choices are:
1) An attached method -- the push method as attached to '[]' -- which
prevents its use as a generic: '[].push.call(a, 3)'.
2) An exophoric function, which enables generic use, but also allows
this-stealing which violates object-capability rules (as explained
further in 803).
3) A "malfunction" -- a frozen object with call, apply, and bind
members that do what these members of the internal exophoric tamed
method would do, but which can't be called directly, either as a
function, as a method, nor as a constructor. (A malfunction is similar
to and inspired by the Valija-level "disfunction", but is frozen and
does not inherit from Disfunction.prototype.)

Prior to rm-bathwater, Caja provided a mixture of #1 and #2. With
rm-bathwater, Cajita provides only #2. This CL changes all
tamed-as-generics to reify instead as malfunctions. This was motived
to fix bug 799 (below), but fixed 803, 418, and 806 as well.

asFirstClass() now rejects exophoric functions, since these should
never be exposed.

reifyIfXo4a() coerces to a first-class value. If given an exophoric
function, it returns a malfunction. Otherwise, it just delegates to
asFirstClass().

Instead of having readPub() call attach() or reifyIfXo4a(), we've
moved this reification logic into fault handlers registered by the
taming methods grantGeneric() and handleGeneric(), since only taming
can now cause the situation. (Cajita code cannot define exophoric
functions.) 




    799: What is Array.prototype.push in Valija?

cajita.isJSONContainer() now says false on prototypical objects
(objects whose .constructor is a function whose .prototype is
themselves). Since prototypical properties are not first-class, this
has no effect on Cajita's semantics, but enables better reuse within
the Cajita implementation.

cajita.getMethodNames(func) has been replaced with
cajita.getProtoPropertyNames(func) and
cajita.getProtoPropertyValue(func,name), giving access to the (coerced
to first-class) elements of func.prototype.

Valija's getShadow() uses these new methods to initialize a shadow
prototype object with a Valija-level emulation of func.prototype.