Make HTML and CSS schema checking more robust by using Name objects to index schemas instead of Strings.  Previously, we were using Strings which had to be compared case-insensitively.  Consistently comparing case-insensitively was hard to maintain, so we started to canonicalize strings, but did not do that consistently either.
This change introduces a Name class with 3 constructors: Name.html,
Name.xml, and Name.css.  They do appropriate case-folding at construct
time, so the only places that need to be checked are where Strings are
converted to Names.

The Name class implements .equals, .hashCode, and Comparable.compareTo
so can be used as a key with the same kinds of Collections a String
can.