Fixes 844, 843, 834, 847
844: privilege escalation:

reifyIfXo4a(), when given a tamed exophoric function (which therefore
may not be first-class), returns a malfunction -- a record with call,
apply, and bind methods that emulate and wrap the safe behavior of the
original function. However, these methods forgot to check whether
their first argument was null or undefined. By passing these
unmofified to the original function, an attacker can cause the
original function to operate instead on the global object. By
replacing such bad this-values with cajita.USELESS, we plug this leak.

834: More broken Cajita callbacks.

xxx



Other changes

* cajita-debugmode.js was replacing cajita.log with an instrumented one,
but forgot to mark it as a simple frozen function.